requirements.txt
asgiref==3.8.1
Django==5.1.5
django-phonenumber-field==8.0.0
sqlparse==0.5.3
typing_extensions==4.12.2
environs==14.1.0
phonenumbers==8.13.53
whitenoise==6.8.2
